fix(shopify): keep server-only onboarding fetchers out of client bundle
`app/routes/app.tsx` uses `validateCompleteOnboarding` on the client and
imported it from `app/utils/onboarding.ts`, which also pulled in
`app/services.server/*` (the pino/AsyncLocalStorage logger + Admin GraphQL
clients). That dragged the whole server graph into the client `onboarding`
chunk, so the browser hit `new AsyncLocalStorage()` (node:async_hooks) and
threw "AsyncLocalStorage is not a constructor", crashing hydration and
leaving the embedded admin screens blank.

rolldown-vite does not enforce the `.server` module boundary that would
normally fail this build, so the leak shipped silently to prod and staging.

Split the module:
- onboarding.server.ts: fetchAllOnboardingData + stepDataFetchers + all
  services.server runtime imports (server-only).
- onboarding.ts: client-safe validators/types only; service types are now
  erased `import type`.
- app.tsx imports fetchAllOnboardingData from onboarding.server.

Verified the rebuilt client bundle no longer contains AsyncLocalStorage,
pino, drizzle, postgres, secrets, or any node: builtin import.
